# contracts/comfyui-sdui.openapi.json
{
  "components": {
...
    "schemas": {
...
      "Problem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "pattern": "^[a-z][a-z0-9_]*$",
            "type": "string"
          },
          "detail": {
            "maxLength": 4000,
            "minLength": 1,
            "type": "string"
          },
          "errors": {
            "items": {
              "additionalProperties": false,
              "properties": {
                "code": {
                  "pattern": "^[a-z][a-z0-9_]*$",
                  "type": "string"
                },
                "detail": {
                  "maxLength": 2000,
                  "minLength": 1,
                  "type": "string"
                },
                "field": {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                }
              },
              "required": [
                "field",
                "code",
                "detail"
              ],
              "type": "object"
            },
            "maxItems": 100,
            "minItems": 1,
            "type": "array"
          },
          "instance": {
            "type": "string"
          },
          "requestId": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "retryable"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local workflow-extensible server-driven UI and durable ComfyUI job API.",
    "title": "ComfyUI SDUI Backend API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/jobs/{jobId}/preview": {
      "get": {
        "operationId": "getJobPreview",
        "parameters": [
          {
            "in": "path",
            "name": "jobId",
            "required": true,
            "schema": {
              "example": "0198aabb-ccdd-7000-8000-0123456789ab",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$/u",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/jpeg": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/png": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Latest bounded transient preview"
          },
          "304": {
            "description": "Preview sequence matches If-None-Match"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid job id"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Job or transient preview does not exist"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unexpected service failure"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Job service is unavailable"
          }
        },
        "tags": [
          "Jobs"
        ]
      }
    },
...
  },
...
}